requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.2.2
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.1
chromadb==0.4.2
click==8.1.6
clickhouse-connect==0.6.8
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
decorator==5.1.1
duckdb==0.8.1
entrypoints==0.4
exceptiongroup==1.1.2
faiss-cpu==1.7.4
Faker==19.2.0
fastapi==0.99.1
favicon==0.7.0
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
htbuilder==0.6.1
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
langchain==0.0.237
langsmith==0.0.10
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mdurl==0.1.2
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.4.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
posthog==3.0.1
protobuf==3.20.3
pulsar-client==3.2.0
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.16.1
pymdown-extensions==10.1
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.15.0
PyPDF2==3.0.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
semver==3.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
st-annotated-text==4.0.0
starlette==0.27.0
streamlit==1.18.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
